fix compile error

+++ velox/dwio/common/SelectiveColumnReader.h
@@ -733,7 +733,7 @@ class SelectiveColumnReader {
 
 template <>
 inline void SelectiveColumnReader::addValue(const std::string_view value) {
-  const uint64_t size = value.size();
+  const uint32_t size = value.size();
   if (size <= StringView::kInlineSize) {
     reinterpret_cast<StringView*>(rawValues_)[numValues_++] =
         StringView(value.data(), size);
